Fix issues with deduplication settings
